pass run_id to subprocesses
