+++ activitysim/cli/run.py
@@ -16,6 +16,7 @@ import numpy as np
 
 from activitysim.core import chunk, config, mem, timing, tracing, workflow
 from activitysim.core.configuration import FileSystem, Settings
+from activitysim.core.run_id import RunId
 
 from activitysim.abm.models.settings_checker import check_model_settings
 
@@ -33,14 +34,6 @@ INJECTABLES = [
     "run_timestamp",
     "run_id",
 ]
-
-class RunId(str):
-    def __new__(cls, x=None):
-        if x is None:
-            return cls(
-                hex(struct.unpack("<Q", struct.pack("<d", time.time()))[0])[-6:].lower()
-            )
-        return super().__new__(cls, x)
 
 
 def add_run_args(parser, multiprocess=True):
@@ -172,7 +165,7 @@ def handle_standard_args(state: workflow.State, args, multiprocess=True):
         # 'configs', 'data', and 'output' folders by default
         os.chdir(args.working_dir)
 
-    inject_arg("run_id", state.run_id)
+    inject_arg("run_id", state.tracing.run_id)
 
     if args.ext:
         for e in args.ext:
@@ -281,7 +274,7 @@ def run(args):
     """
 
     state = workflow.State()
-    state.run_id = RunId()
+    _init_run_id = state.tracing.run_id
 
     # register abm steps and other abm-specific injectables
     # by default, assume we are running activitysim.abm
