+++ activitysim/core/configuration/top.py
@@ -791,10 +791,3 @@ class Settings(PydanticBase, extra="allow", validate_assignment=True):
         except AttributeError:
             return self.other_settings.get(attr)
 
-class RunId(str):
-    def __new__(cls, x=None):
-        if x is None:
-            return cls(
-                hex(struct.unpack("<Q", struct.pack("<d", time.time()))[0])[-6:].lower()
-            )
-        return super().__new__(cls, x)
