+++ activitysim/core/workflow/tracing.py
@@ -20,6 +20,7 @@ from activitysim.core import tracing
 from activitysim.core.test import assert_equal, assert_frame_substantively_equal
 from activitysim.core.workflow.accessor import FromState, StateAccessor
 from activitysim.core.exceptions import TableSlicingError
+from activitysim.core.run_id import RunId
 
 logger = logging.getLogger(__name__)
 
@@ -42,7 +43,7 @@ class Tracing(StateAccessor):
     traceable_tables: list[str] = FromState(default_value=DEFAULT_TRACEABLE_TABLES)
     traceable_table_ids: dict[str, Sequence] = FromState(default_init=True)
     traceable_table_indexes: dict[str, str] = FromState(default_init=True)
-    # run_id: RunId = FromState(default_init=True)
+    run_id: RunId = FromState(default_init=True)
 
     @property
     def validation_directory(self) -> Path | None:
@@ -241,7 +242,7 @@ class Tracing(StateAccessor):
             file_name = "%s.%s" % (file_name, CSV_FILE_TYPE)
 
         file_path = self._obj.filesystem.get_trace_file_path(
-            file_name, tail=self._obj.run_id
+            file_name, tail=self.run_id
         )
 
         if os.name == "nt":
@@ -367,7 +368,7 @@ class Tracing(StateAccessor):
 
                         that_blob = read_csv_as_list_of_lists(that_path)
                         this_path = self._obj.filesystem.get_trace_file_path(
-                            label, tail=self._obj.run_id, file_type="csv"
+                            label, tail=self.run_id, file_type="csv"
                         )
                         this_blob = read_csv_as_list_of_lists(this_path)
 
@@ -399,7 +400,7 @@ class Tracing(StateAccessor):
                         that_df = pd.read_csv(that_path)
                         # check against the file we just wrote
                         this_path = self._obj.filesystem.get_trace_file_path(
-                            label, tail=self._obj.run_id, file_type="csv"
+                            label, tail=self.run_id, file_type="csv"
                         )
                         this_df = pd.read_csv(this_path)
                         assert_frame_substantively_equal(this_df, that_df)
@@ -441,7 +442,7 @@ class Tracing(StateAccessor):
         # write out the raw dataframe
 
         file_path = self._obj.filesystem.get_trace_file_path(
-            "%s.raw.csv" % label, tail=self._obj.run_id
+            "%s.raw.csv" % label, tail=self.run_id
         )
         trace_results.to_csv(file_path, mode="a", index=True, header=True)
